refactor delete to be subscription
