+++ apis/api-gateway/schema.graphql
@@ -425,8 +425,6 @@ type Mutation @join__type(graph: API_ANALYTICS)  @join__type(graph: API_JOURNEYS
     """
     timezone: String
   ): JourneyVisitorGoogleSheetExportResult! @join__field(graph: API_JOURNEYS_MODERN) 
-  userDeleteJourneysCheck(userId: String!) : UserDeleteJourneysCheckResult! @join__field(graph: API_JOURNEYS_MODERN) 
-  userDeleteJourneysConfirm(userId: String!) : UserDeleteJourneysConfirmResult! @join__field(graph: API_JOURNEYS_MODERN) 
   audioPreviewCreate(input: MutationAudioPreviewCreateInput!) : AudioPreview! @join__field(graph: API_LANGUAGES) 
   audioPreviewUpdate(input: MutationAudioPreviewUpdateInput!) : AudioPreview! @join__field(graph: API_LANGUAGES) 
   audioPreviewDelete(languageId: ID!) : AudioPreview! @join__field(graph: API_LANGUAGES) 
@@ -557,7 +555,6 @@ type Mutation @join__type(graph: API_ANALYTICS)  @join__type(graph: API_JOURNEYS
   createVerificationRequest(input: CreateVerificationRequestInput) : Boolean @join__field(graph: API_USERS) 
   validateEmail(email: String!, token: String!) : AuthenticatedUser @join__field(graph: API_USERS) 
   userDeleteCheck(idType: UserDeleteIdType!, id: String!) : UserDeleteCheckResult! @join__field(graph: API_USERS) 
-  userDeleteConfirm(idType: UserDeleteIdType!, id: String!) : UserDeleteResult! @join__field(graph: API_USERS) 
 }
 
 type MutationSiteCreateSuccess @join__type(graph: API_ANALYTICS)  {
@@ -2527,8 +2524,9 @@ type PlausibleStatsResponse @join__type(graph: API_JOURNEYS_MODERN)  {
   timeOnPage: Float
 }
 
-type Subscription @join__type(graph: API_JOURNEYS_MODERN)  {
-  journeyAiTranslateCreateSubscription(input: JourneyAiTranslateInput!) : JourneyAiTranslateProgress!
+type Subscription @join__type(graph: API_JOURNEYS_MODERN)  @join__type(graph: API_USERS)  {
+  journeyAiTranslateCreateSubscription(input: JourneyAiTranslateInput!) : JourneyAiTranslateProgress! @join__field(graph: API_JOURNEYS_MODERN) 
+  userDeleteConfirm(idType: UserDeleteIdType!, id: String!) : UserDeleteConfirmProgress! @join__field(graph: API_USERS) 
 }
 
 type TemplateFamilyStatsAggregateResponse @join__type(graph: API_JOURNEYS_MODERN)  {
@@ -2548,31 +2546,6 @@ type TemplateFamilyStatsBreakdownResponse @join__type(graph: API_JOURNEYS_MODERN
 type TemplateFamilyStatsEventResponse @join__type(graph: API_JOURNEYS_MODERN)  {
   event: String!
   visitors: Int!
-}
-
-type UserDeleteJourneysCheckResult @join__type(graph: API_JOURNEYS_MODERN)  {
-  journeysToDelete: Int!
-  journeysToTransfer: Int!
-  journeysToRemove: Int!
-  teamsToDelete: Int!
-  teamsToTransfer: Int!
-  teamsToRemove: Int!
-  logs: [UserDeleteJourneysLogEntry!]!
-}
-
-type UserDeleteJourneysConfirmResult @join__type(graph: API_JOURNEYS_MODERN)  {
-  success: Boolean!
-  deletedJourneyIds: [String!]!
-  deletedTeamIds: [String!]!
-  deletedUserJourneyIds: [String!]!
-  deletedUserTeamIds: [String!]!
-  logs: [UserDeleteJourneysLogEntry!]!
-}
-
-type UserDeleteJourneysLogEntry @join__type(graph: API_JOURNEYS_MODERN)  {
-  message: String!
-  level: String!
-  timestamp: String!
 }
 
 type YouTube @join__type(graph: API_JOURNEYS_MODERN, key: "id primaryLanguageId", extension: true)  {
@@ -3330,15 +3303,16 @@ type UserDeleteCheckResult @join__type(graph: API_USERS)  {
   logs: [UserDeleteLogEntry!]!
 }
 
+type UserDeleteConfirmProgress @join__type(graph: API_USERS)  {
+  log: UserDeleteLogEntry!
+  done: Boolean!
+  success: Boolean
+}
+
 type UserDeleteLogEntry @join__type(graph: API_USERS)  {
   message: String!
   level: String!
   timestamp: String!
-}
-
-type UserDeleteResult @join__type(graph: API_USERS)  {
-  success: Boolean!
-  logs: [UserDeleteLogEntry!]!
 }
 
 interface BaseError @join__type(graph: API_ANALYTICS)  @join__type(graph: API_MEDIA)  {
