+++ apis/api-journeys-modern/schema.graphql
@@ -1406,8 +1406,6 @@ type Mutation {
     """
     timezone: String
   ): JourneyVisitorGoogleSheetExportResult!
-  userDeleteJourneysCheck(userId: String!): UserDeleteJourneysCheckResult!
-  userDeleteJourneysConfirm(userId: String!): UserDeleteJourneysConfirmResult!
 }
 
 input MutationJourneyLanguageAiDetectInput {
@@ -2288,31 +2286,6 @@ type UserAgent
   browser: Browser!
   device: Device!
   os: OperatingSystem!
-}
-
-type UserDeleteJourneysCheckResult {
-  journeysToDelete: Int!
-  journeysToTransfer: Int!
-  journeysToRemove: Int!
-  teamsToDelete: Int!
-  teamsToTransfer: Int!
-  teamsToRemove: Int!
-  logs: [UserDeleteJourneysLogEntry!]!
-}
-
-type UserDeleteJourneysConfirmResult {
-  success: Boolean!
-  deletedJourneyIds: [String!]!
-  deletedTeamIds: [String!]!
-  deletedUserJourneyIds: [String!]!
-  deletedUserTeamIds: [String!]!
-  logs: [UserDeleteJourneysLogEntry!]!
-}
-
-type UserDeleteJourneysLogEntry {
-  message: String!
-  level: String!
-  timestamp: String!
 }
 
 type UserInvite
