+++ apis/api-journeys-modern/src/schema/schema.ts
@@ -25,7 +25,6 @@ import './plausible'
 import './qrCode'
 import './team'
 import './user'
-import './userDelete'
 import './userInvite'
 import './userJourney'
 import './userRole'
