--- apis/api-journeys-modern/src/schema/userDelete/index.ts
@@ -1,2 +1,0 @@
-import './userDeleteJourneysCheck'
-import './userDeleteJourneysConfirm'
