+++ apis/api-users/infrastructure/locals.tf
@@ -11,6 +11,7 @@ locals {
     "JESUS_FILM_PROJECT_VERIFY_URL",
     "JOURNEYS_ADMIN_URL",
     "NAT_ADDRESSES",
+    "PG_DATABASE_URL_JOURNEYS",
     "PG_DATABASE_URL_USERS",
     "REDIS_PORT",
     "REDIS_URL",
