+++ apis/api-users/src/schema/builder.ts
@@ -112,3 +112,4 @@ export const builder = new SchemaBuilder<{
 
 builder.queryType({})
 builder.mutationType({})
+builder.subscriptionType({})
